match_score: Float -> Integer (Phase 3.8)
Closes the last audit item. match_score always holds a rounded 0–100 integer
(matching_service stores round(...)); the column was Float. Change it to Integer
for honest typing.

- app/models/property.py: match_score column Float -> Integer.
- Alembic revision c3a4d5e6f7a8 with postgresql_using to cast existing values
  (double precision -> integer isn't implicit). Verified on Postgres 16: an
  existing 87.0 becomes 87, upgrade->downgrade->upgrade is reversible, and
  alembic check reports no drift.
- BDD step now assigns int(score) rather than float(score).

Suite: 150 pytest (95% cov), 9 features / 55 scenarios behave, ruff clean.

+++ backend/app/models/property.py
@@ -66,7 +66,8 @@ class Property(Base, TimestampMixin):
     # Investment / scoring fields used by comparison.
     estimated_appreciation_pct: Mapped[Optional[float]] = mapped_column(Float)
     rental_yield: Mapped[Optional[str]] = mapped_column(String(32))
-    match_score: Mapped[Optional[float]] = mapped_column(Float)
+    # Stored as an integer 0–100 (matching_service rounds the weighted score).
+    match_score: Mapped[Optional[int]] = mapped_column(Integer)
 
     notes: Mapped[Optional[str]] = mapped_column(Text)
